ctx: add file path of worker to ctx
this adds the filepath of the inititated worker to the context.
this is useful for several things, for me it is the most flexible
way to detect the file name of the corresponding `loc.api` api
filehandler via `getApiConf`.

This way we can remove `getApiConf` from most workers. The method
just needs extension in special cases. `wrk-api` would receive
the current worker, construct the correpsonding loc.api filename
from it, and return it in `getApiConf`.